ytdl_hook: make safe url match non-greedy
The previous regexp was greedy and used to eat up more than the protocol
if the rest of the url contained at least another non encoded `://`.
For example, trying to match "https://fqdn/?redirect=https://redir"
would wrongly qualify the whole "https://fqdn/?redirect=https" as the
protocol.

Technically, this pattern would be "^(..-)://" to preserve the previous
intention, but the result is the same with "^(.-)://".